requirements.txt
proxmoxer
django
djangorestframework
django-tables2
 